backend/ml/requirements.txt
# ML/AI Dependencies for Careerak Recommendation System
# نظام التوصيات الذكية - المتطلبات

# Core ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# NLP Libraries
spacy==3.7.2
nltk==3.8.1

# Arabic NLP Support
camel-tools==1.5.2
pyarabic==0.6.15

# Text Processing
sentence-transformers==2.2.2
transformers==4.36.2

# PDF/Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Feature Engineering
scipy==1.11.4
joblib==1.3.2

# Model Persistence
pickle-mixin==1.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: Deep Learning (if needed later)
# torch==2.1.2
# tensorflow==2.15.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
